BROKEN BUILD SERIES: Upstream 2c90fe2 Dec 5, 2010

+++ qemu/savevm.c
@@ -62,7 +62,6 @@
 #endif
 
 #ifdef _WIN32
-#include <WinSock2.h>
 #include <windows.h>
 #include <malloc.h>
 #include <sys/timeb.h>
@@ -612,7 +611,7 @@ int qemu_file_rate_limit(QEMUFile *f)
     return 0;
 }
 
-size_t qemu_file_get_rate_limit(QEMUFile *f)
+int64_t qemu_file_get_rate_limit(QEMUFile *f)
 {
     if (f->get_rate_limit)
         return f->get_rate_limit(f->opaque);
@@ -620,7 +619,7 @@ size_t qemu_file_get_rate_limit(QEMUFile *f)
     return 0;
 }
 
-size_t qemu_file_set_rate_limit(QEMUFile *f, size_t new_rate)
+int64_t qemu_file_set_rate_limit(QEMUFile *f, int64_t new_rate)
 {
     /* any failed or completed migration keeps its state to allow probing of
      * migration data, but has no associated file anymore */
